Move releases to master branch, use `npm version`
[Issue 291] reminded me how much of a pain it is to prepare and publish
a new release, so I thought I'd try out something different, inspired by
[webtorrent]. Instead of having a separate release branch, let's just
keep releases on master. This avoids all the extra git stuff required to
do it the old way, while still supporting bower users by checking in
`dist/` files upon release.

With these changes, preparing a release is as simple as

    npm version patch # patch/minor/major/etc

Using the [scripts] in package.json, this will:

* Remove old `dist/` files
* Reinstall `node_modules` from scratch
* Do a clean build and test
* Add new `dist/` files
* Bump version numbers in `package.json` and `bower.json`
* Commit and tag the release

This result will be a single tagged commit with all the changes in
`dist/`, `package.json`, and `bower.json`.

After that, all that's needed to publish the release is:

    npm publish # for npm
    git push --follow-tags # for bower

`npm publish` creates a new commit that removes the `dist/` directory,
so that it can remain ignored until the next release.

[Issue 291]: https://github.com/onsip/SIP.js/pull/291#issuecomment-208897713
[webtorrent]: https://github.com/feross/webtorrent
[scripts]: https://docs.npmjs.com/cli/version